Fixed a bug in my code's logic when inferring maps from Python dictionaries imported from JSON files
Fixed a bug in my code's logic when inferring maps from Python dictionaries imported from JSON files.  I should only error out if none of the types are found, so the blocks should all be elif clauses (not if clauses) with a final else that leads to an error only if none of the preceding types are found.